remove assign lyric layout logic in the legacy beatmap process generator. also change the class naming.
